admin-panel-scratch-3/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pandas==2.2.0
openpyxl==3.1.2
aiofiles==23.2.1
google-cloud-storage==2.14.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
mongomock-motor==0.0.29
aiosmtplib==3.0.1
email-validator==2.1.0.post1
